[FIA] modifiche interfaccia grafica messaggio, aggiunta visualizzazione dettagliata contenuto consigliato dal chatbot, modifiche varie per miglioramento prestazioni chatbot

+++ app/src/main/java/it/unisa/ilike/contenuti/storage/FilmBean.java
@@ -1,6 +1,7 @@
 package it.unisa.ilike.contenuti.storage;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * Questa classe contiene gli attributi e i metodi di utilità relativi ai film
@@ -27,7 +28,7 @@ public class FilmBean extends ContenutoBean  implements Serializable {
      * @param regista è il regista del film.
      * @param attori è l'elenco degli attori del film.
      */
-    public FilmBean(int id, String titolo, String descrizione, String categoria, Double annoRilascio, int durata, String paese, String regista, String attori, float valutazioneMedia) {
+    public FilmBean(int id, String titolo, String descrizione, String categoria, BigDecimal annoRilascio, int durata, String paese, String regista, String attori, float valutazioneMedia) {
         super(id, titolo, descrizione, categoria, valutazioneMedia);
         this.annoRilascio = annoRilascio;
         this.durata = durata;
@@ -36,7 +37,7 @@ public class FilmBean extends ContenutoBean  implements Serializable {
         this.attori = attori;
     }
 
-    public FilmBean (int id, String titolo, Double annoRilascio, String categoria,String paese, String regista, String descrizione){
+    public FilmBean (int id, String titolo, BigDecimal annoRilascio, String categoria, String paese, String regista, String descrizione){
         super(id, titolo, descrizione, categoria, 0);
         this.setId(id);
         this.setTitolo(titolo);
@@ -53,7 +54,7 @@ public class FilmBean extends ContenutoBean  implements Serializable {
      * Questo metodo restituisce l'anno di rilascio del film
      * @return anno di rilascio
      */
-    public Double getAnnoRilascio() {
+    public BigDecimal getAnnoRilascio() {
         return annoRilascio;
     }
 
@@ -61,7 +62,7 @@ public class FilmBean extends ContenutoBean  implements Serializable {
      * Questo metodo imposta l'anno di rilascio del film
      * @param annoRilascio
      */
-    public void setAnnoRilascio(Double annoRilascio) {
+    public void setAnnoRilascio(BigDecimal annoRilascio) {
         this.annoRilascio = annoRilascio;
     }
 
@@ -133,7 +134,7 @@ public class FilmBean extends ContenutoBean  implements Serializable {
         return ""+this.annoRilascio;
     }
 
-    private Double annoRilascio;
+    private BigDecimal annoRilascio;
     private int durata;
     private String paese;
     private String regista;
